Django Developer Task/task1/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
django-rest==0.8.7
djangorestframework==3.13.1
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
